Take Switch + Steam hidraw from #41 and the 2.15:1 docked touch pad.
Keep main's remappable KBM-as-XInput (do not write friend keyboards
into PCSX2.ini) and #44 landscape side-mode. Drop emulator_keymap.

+++ crates/client/src/steam_reader.rs
@@ -62,7 +62,7 @@ mod linux_impl {
 
     /// `HIDIOCSFEATURE(len)` = `_IOWR('H', 0x06, len)`.
     fn hidio_sfeature(len: usize) -> usize {
-        (3u64 << 30) | ((len as u64) << 16) | (0x48 << 8) | 0x06
+        ((3u64 << 30) | ((len as u64) << 16) | (0x48 << 8) | 0x06) as usize
     }
 
     pub struct SteamReader {
